Backend/python/chatwithdocument-python/requirements.txt
Flask 
requests 
pdfplumber 
pandas 
openpyxl 
Pillow
